Removing coop interfaces
