--- library/include/rocwmma/internal/coop_io_bearer_impl.hpp
@@ -1,91 +1,0 @@
-/*******************************************************************************
- *
- * MIT License
- *
- * Copyright (C) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- *******************************************************************************/
-#ifndef ROCWMMA_COOP_IO_BEARER_IMPL_HPP
-#define ROCWMMA_COOP_IO_BEARER_IMPL_HPP
-
-#include "io_bearer.hpp"
-#include "io_scheduler.hpp"
-#include "layout/matrix_coop_layout.hpp"
-#include "utility/math.hpp"
-#include "vector.hpp"
-
-namespace rocwmma
-{
-
-#define IOBearerTypesDecl                 \
-    class DataLayout, class MatrixLayout, \
-        template <typename, uint32_t>     \
-        class BearerPolicy, class BoundsCtrl, class Scheduler
-
-#define IOBearerTypesImpl DataLayout, MatrixLayout, BearerPolicy, BoundsCtrl, Scheduler
-
-    template <IOBearerTypesDecl>
-    template <typename BufferT, typename ExternDataT>
-    ROCWMMA_DEVICE /* static */ inline void
-        IOBearer<IOBearerTypesImpl>::exec(BufferT&& buffer, ExternDataT* dataPtr, uint32_t ldm)
-    {
-        // Traits
-        using SchedulerTraits = scheduler_traits<Scheduler>;
-
-        // Sanity check
-        static_assert(SchedulerTraits::is_valid, "Invalid scheduler");
-
-        // Cooperative
-        if constexpr(SchedulerTraits::is_cooperative)
-        {
-            // Waves filter
-            if(!MatrixLayout::waveEnabler(Scheduler::waveIndex(), Scheduler::waveCount()))
-            {
-                return;
-            }
-
-            // We need to shrink the buffer for the unroll
-            using BuffTraits = VecTraits<decay_t<BufferT>>;
-            using DataT      = typename BuffTraits::DataT;
-
-            constexpr auto BuffSize
-                = reduce_mult(MatrixLayout::strideCounts(Scheduler::waveCount()))
-                  * BaseImpl::TransactionSize;
-            using ReducedBufferT
-                = conditional_t<is_const_v<BufferT>,
-                                typename BuffTraits::template VecT<DataT, BuffSize> const,
-                                typename BuffTraits::template VecT<DataT, BuffSize>>;
-
-            BaseImpl::exec(forward<ReducedBufferT>(buffer), dataPtr, ldm);
-        }
-        // Non-cooperative
-        else
-        {
-            BaseImpl::exec(forward<BufferT>(buffer), dataPtr, ldm);
-        }
-    }
-
-#undef IOBearerTypesDecl
-#undef IOBearerTypesImpl
-
-} // namespace rocwmma
-
-#endif // ROCWMMA_COOP_IO_BEARER_IMPL_HPP
