--- library/include/rocwmma/internal/coop_store.hpp
@@ -1,64 +1,0 @@
-/*******************************************************************************
- *
- * MIT License
- *
- * Copyright (C) 2021-2025 Advanced Micro Devices, Inc. All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- *******************************************************************************/
-#ifndef ROCWMMA_COOP_STORE_HPP
-#define ROCWMMA_COOP_STORE_HPP
-
-#include "coop_io_bearer.hpp"
-#include "io_bounds_ctrl.hpp"
-#include "layout/matrix_coop_layout_impl.hpp"
-#include "opaque_store.hpp"
-
-namespace rocwmma
-{
-    // Storing interface
-    template <class DataLayout,
-              class MatrixLayout,
-              class BoundsCtrl = IOBoundsCtrl::Default,
-              class Scheduler  = IOScheduler::Default>
-    struct CooperativeStore : public IOBearer<DataLayout,
-                                              MatrixLayout,
-                                              detail::OpaqueStoreBearer,
-                                              BoundsCtrl,
-                                              Scheduler>
-    {
-    private:
-        using Base
-            = IOBearer<DataLayout, MatrixLayout, detail::OpaqueStoreBearer, BoundsCtrl, Scheduler>;
-
-        // Don't expose the base implementation, we change arg forwarding order
-        using Base::exec;
-
-    public:
-        template <typename DataT, typename BufferT>
-        ROCWMMA_DEVICE static void exec(DataT* dataPtr, BufferT&& buff, uint32_t ldm)
-        {
-            Base::exec(forward<BufferT>(buff), dataPtr, ldm);
-        }
-    };
-
-} // namespace rocwmma
-
-#endif // ROCWMMA_COOP_STORE_HPP
